Prevent double toggling of nav pane when game selected
Clicks on the game select buttons in the left-hand nav pane
were being handled by two event handlers, both of which were
toggling the open/closed state of the nav pane.

By calling preventDefault() on the event in the handlers for
the buttons, we prevent the event from being propagated
up to the nav pane itself.